backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
python-dotenv>=1.0
PyMySQL>=1.0
Flask-Cors>=3.0
gunicorn>=20.1.0
pytest>=7.0
pytest-mock>=3.5
PyJWT>=2.8
bcrypt>=4.0
Flask-JWT-Extended>=4.0.0

